- clean up contracts folder else it compiles only one contract and then gives error for 2nd - add check before removing the cdt container - update dockerfile version in config file to rebuild

+++ packages/api-eosio-compiler/docker-eosio-cdt/remove_eosio_cdt_docker.sh
@@ -9,4 +9,6 @@ if [ -f "./config.file.local" ]; then
   source ./config.file.local
 fi
 
-docker rm -fv $CDT_CONTAINER_NAME
+if [ "$(docker ps -q -f name=$CDT_CONTAINER_NAME)" ]; then
+  docker rm -fv $CDT_CONTAINER_NAME
+fi
